Removed "LastLogin" and "LastUpdated" from User and Node
I'm choosing to remove these because they are not yet functional and there are more important features to work on.

LastLogin could be accomplished with another update procedure that is called whenever a user logs in.

LastUpdated should really have been LastSignal. This would be updated any time a row is added to the NodeData table and would tell us the last time each node has sent us data. It could be accomplished with a trigger that fires whenever values are inserted into the NodeData table. The trigger would then set LastSignal in the Node table to the current time, GETDATE().